Moving files and updating namespaces.
